Refactor EmailSettings
